calendar-ai-bot/requirements.txt
# Core dependencies
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
pytz>=2023.3

# LLM integration
groq>=0.4.0
openai>=1.3.0

# Utilities
python-dateutil>=2.8.2
pydantic>=2.4.0
cachetools>=5.3.1

# Logging and configuration
structlog>=23.1.0

# Type hints
typing-extensions>=4.7.1

# Performance
asyncio>=3.4.3
